+++ packages/snap-preact-components/src/components/Organisms/FilterSummary/FilterSummary.test.tsx
@@ -5,6 +5,7 @@ import { ThemeProvider } from '../../../providers';
 
 import { FilterSummary } from './FilterSummary';
 import { filters } from '../../../mocks/store';
+import userEvent from '@testing-library/user-event';
 
 describe('FilterSummary Component', () => {
 	it('renders with filter list', () => {
@@ -52,6 +53,22 @@ describe('FilterSummary Component', () => {
 		expect(title).toHaveTextContent('you clicked these earlier');
 	});
 
+	it('renders with specified icons', async () => {
+		const args = {
+			filters: filters,
+			clearAllIcon: 'circle',
+			filterIcon: 'check',
+		};
+
+		const rendered = render(<FilterSummary {...args} />);
+
+		const filterIcon = rendered.container.querySelector('.ss__filter-summary .ss__filter .ss__icon');
+		const clearAllIcon = rendered.container.querySelector('.ss__filter-summary__clear-all .ss__icon');
+
+		expect(filterIcon).toHaveClass(`ss__icon--${args.filterIcon}`);
+		expect(clearAllIcon).toHaveClass(`ss__icon--${args.clearAllIcon}`);
+	});
+
 	it('can hide the facet label', () => {
 		const rendered = render(<FilterSummary filters={filters} hideFacetLabel={true} />);
 		const facetLabel = rendered.container.querySelector('.ss__filter__label');
@@ -63,6 +80,71 @@ describe('FilterSummary Component', () => {
 		const FilterElement = rendered.container.querySelector('.ss__filter-summary');
 
 		expect(FilterElement).not.toBeInTheDocument();
+	});
+
+	it('renders with custom seperator', () => {
+		const sep = '>>>';
+		const rendered = render(<FilterSummary filters={filters} separator={sep} />);
+		const FilterElement = rendered.container.querySelector('.ss__filter-summary');
+
+		expect(FilterElement).toBeInTheDocument();
+		const seperatorElem = rendered.container.querySelector('.ss__filter__label__separator');
+
+		expect(seperatorElem).toHaveTextContent(sep);
+	});
+
+	it('renders with custom onclick func', () => {
+		const onclickfunc = jest.fn();
+
+		const rendered = render(<FilterSummary filters={filters} onClick={onclickfunc} />);
+		const FilterSumElement = rendered.container.querySelector('.ss__filter-summary');
+
+		expect(FilterSumElement).toBeInTheDocument();
+		const filter = rendered.container.querySelector('.ss__filter-summary .ss__filter');
+
+		expect(filter).toBeInTheDocument();
+
+		userEvent.click(filter);
+
+		expect(onclickfunc).toHaveBeenCalled();
+	});
+
+	it('renders with custom on clear all click func', () => {
+		const onclickfunc = jest.fn();
+
+		const rendered = render(<FilterSummary filters={filters} onClearAllClick={onclickfunc} />);
+		const FilterSumElement = rendered.container.querySelector('.ss__filter-summary');
+
+		expect(FilterSumElement).toBeInTheDocument();
+		const filter = rendered.container.querySelector('.ss__filter-summary .ss__filter-summary__clear-all');
+
+		expect(filter).toBeInTheDocument();
+
+		userEvent.click(filter);
+
+		expect(onclickfunc).toHaveBeenCalled();
+	});
+
+	it('renders with classname', () => {
+		const args = {
+			className: 'classy',
+		};
+
+		const rendered = render(<FilterSummary filters={filters} {...args} />);
+
+		const facetsElement = rendered.container.querySelector('.ss__filter-summary');
+		expect(facetsElement).toHaveClass(args.className);
+	});
+
+	it('disables styles', () => {
+		const args = {
+			disableStyles: true,
+		};
+
+		const rendered = render(<FilterSummary filters={filters} {...args} />);
+
+		const facetsElement = rendered.container.querySelector('.ss__filter-summary');
+		expect(facetsElement.classList).toHaveLength(1);
 	});
 });